requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
pypdf
python-multipart
beautifulsoup4
PyMuPDF
